remove unnecessary deps and discover folder, file on domain
